Built the bones of the Knight and the Hound
I built the framework that I'll use to complete the knight and the hound, as well as made a new enemy behavior script that stores several enemy behaviors in an array such that the enemy will chose a random behavior to execute. None of the other behaviors that I've created actually work yet, but I did leave instructions to my future self on how to build it.