Made a prototype Security Checkpoint
I made a placeholder security checkpoint level with the idea being searching for each door to get towards the inner most hallway in able to complete the level. There are a few different enemy types depending on how far deep in the security checks the player goes. The area is well populated with lamps, as well as having an open ceiling. It is still missing metal gates, scanners, security terminals, and spotlights.